Moving subpopulations and consent models into their own package (consents now related to subpopulations, not studies
