+++ app/org/sagebionetworks/bridge/models/subpopulations/StudyConsentView.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.bridge.models.studies;
+package org.sagebionetworks.bridge.models.subpopulations;
 
 import static com.google.common.base.Preconditions.checkNotNull;
