+++ app/org/sagebionetworks/bridge/models/subpopulations/Subpopulation.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.bridge.models.studies;
+package org.sagebionetworks.bridge.models.subpopulations;
 
 import java.util.Set;
